chore: move method to avoid circular deps
