+++ packages/commons/src/events/scopes.ts
@@ -21,6 +21,8 @@ import {
   ActionType,
   DisplayableAction
 } from './ActionType'
+import { EventIndex } from './EventIndex'
+import { CreatedAction } from './ActionDocument'
 
 type AlwaysAllowed = null
 
@@ -109,4 +111,31 @@ export function isActionInScope(
     allowedConfigurableScopes,
     eventType
   )
+}
+
+/**
+ * A shared utility to check if the user can read a record.
+ * This will be removed in 1.10 and implemented by scopes.
+ *
+ * In order for us to limit the usage of 'record.read' scope, we allow users to view records they have created on system-level.
+ *
+ * @deprecated - Will be removed in 1.10
+ */
+export function canUserReadRecord(
+  event: EventIndex | CreatedAction,
+  {
+    userId,
+    scopes
+  }: {
+    userId: string
+    scopes: string[]
+  }
+) {
+  const createdByUser = event.createdBy === userId
+
+  if (createdByUser) {
+    return true
+  }
+
+  return isActionInScope(scopes, ActionType.READ, event.type)
 }
