+++ packages/commons/src/scopes.ts
@@ -539,30 +539,3 @@ export const ActionScopes = z.union([
 ])
 
 export type ActionScopes = z.infer<typeof ActionScopes>
-
-/**
- * A shared utility to check if the user can read a record.
- * This will be removed in 1.10 and implemented by scopes.
- *
- * In order for us to limit the usage of 'record.read' scope, we allow users to view records they have created on system-level.
- *
- * @deprecated - Will be removed in 1.10
- */
-export function canUserReadRecord(
-  event: EventIndex | CreatedAction,
-  {
-    userId,
-    scopes
-  }: {
-    userId: string
-    scopes: string[]
-  }
-) {
-  const createdByUser = event.createdBy === userId
-
-  if (createdByUser) {
-    return true
-  }
-
-  return isActionInScope(scopes, ActionType.READ, event.type)
-}
